requirements.txt
python-dotenv==1.0.0
python-telegram-bot==20.7
openai==1.3.0
pdfplumber==0.10.0
python-docx==0.8.11